INC-224 refactor: extract ProfileModal into standalone component
