+++ src/components/pagesComponents/profile/profileModal/index.ts
@@ -1,0 +1,1 @@
+export * from './ProfileModal'
